+++ libs/react/src/components/organisms/complianceWizardLair/complianceWizardLair.tsx
@@ -52,7 +52,7 @@ export const ComplianceWizardLair = (props: PropsWithChildren<ComplianceWizardLa
               label={'Automate'}
               variant={ButtonTypes.primary}
               onClick={() => {
-                setCurrentStep('processing');
+                setCurrentStep('automate');
               }}
             />
             <BaseButton